Update giveaway.js
changed the ouput of all of the editList function to be true or false, if the output is true the code will post the message if not an error will be prompted to the user.

edited the checkMon function so that it would check the status of all of the pokemon stored in the list and if the argument the user is trying to pass are valid.